delete some unused code
