--- frontend/react/src/components/sections/basicinfoapi/backend-json-section-0.json
@@ -1,111 +1,0 @@
-{
-  "section": {
-    "year": 2020,
-    "state": null,
-    "valid": null,
-    "ordinal": 0,
-    "id": "2020-00",
-    "type": "section",
-    "title": "Carts Basic Info",
-    "subsections": [
-      {
-        "type": "subsection",
-        "id": "2020-00-a",
-        "parts": [
-          {
-            "id": "2020-00-a-01",
-            "type": "part",
-            "title": "Welcome!",
-            "questions": [
-              {
-                "id": "2020-00-a-01-01",
-                "label": "State or territory name:",
-                "type": "text",
-                "answer": {
-                  "prepopulated": true,
-                  "readonly": true,
-                  "entry": null
-                }
-              },
-              {
-                "id": "2020-00-a-01-02",
-                "label": "Program type:",
-                "type": "checkbox",
-                "answer": {
-                  "prepopulated": true,
-                  "readonly": true,
-                  "entry": null,
-                  "options": {
-                    "Combination (Medicaid Expansion CHIP and Separate CHIP)": "combo",
-                    "Medicaid Expansion CHIP only": "medicaid_exp_chip",
-                    "Separate CHIP only": "separate_chip"
-                  }
-                }
-              },
-              {
-                "id": "2020-00-a-01-03",
-                "label": "CHIP program name(s): ",
-                "type": "text_medium",
-                "answer": {
-                  "prepopulated": true,
-                  "readonly": true,
-                  "entry": null
-                }
-              }
-            ],
-            "helpdesk": "If any of the above information is incorrect, contact CARTS Help Desk."
-          },
-          {
-            "id": "2020-00-a-02",
-            "type": "part",
-            "title": "Contact information",
-            "text": "Who should we contact if we have any questions about your report?",
-            "questions": [
-              {
-                "id": "2020-00-a-02-01",
-                "label": "Contact name:",
-                "type": "text",
-                "answer": {
-                  "entry": null
-                }
-              },
-              {
-                "id": "2020-00-a-02-02",
-                "label": "Job title:",
-                "type": "text",
-                "answer": {
-                  "entry": null
-                }
-              },
-              {
-                "id": "2020-00-a-02-03",
-                "label": "Email:",
-                "type": "email",
-                "answer": {
-                  "entry": null
-                }
-              },
-              {
-                "id": "2020-00-a-02-04",
-                "label": "Full mailing address:",
-                "hint": "Include city, state and zip code",
-                "type": "mailing_address",
-                "answer": {
-                  "entry": null
-                }
-              },
-              {
-                "id": "2020-00-a-02-05",
-                "label": "Phone number:",
-                "type": "phone_number",
-                "answer": {
-                  "entry": null
-                }
-              }
-            ]
-          }
-        ]
-      }
-    ]
-  }
-}
