--- frontend/react/src/components/sections/section1api/basic.json
@@ -1,201 +1,0 @@
-[
-  {
-    "id": "1",
-    "uuid": "a98f7f6d-1ea9-407a-b175-d21015450264e",
-    "type": "textfield",
-    "properties": {
-      "label": "Briefly describe your goal.",
-      "hint": "For example: Enroll 75% of eligible children in the CHIP program.",
-      "name": "goal_description",
-      "multiline": true,
-      "title": "Define your numerator you’re measuring"
-    }
-  }, 
-  {
-    "id": "2",
-    "uuid": "9020d15c-3272-4e11-9616-0c8a0dfff556",
-    "type": "choicelist",
-    "properties": {
-      "label": "What type of goal is it?",
-      "hint": "For example: Enroll 75% of eligible children in the CHIP program.",
-      "name": "goal_type",
-      "children": [
-        {
-          "type": "radio",
-          "properties": {
-            "defaultChecked": true,
-            "label": "New goal",
-            "value": "new"
-          }
-        },
-        {
-          "type": "radio",
-          "properties": {
-            "label": "Continuing goal", 
-            "value": "continuing"
-          }
-        },
-        {
-          "type": "radio",
-          "properties": {
-            "label": "Discontinued goal", 
-            "value": "discontinued"
-          }
-        }
-      ]
-    }
-  }, 
-  {
-    "id": "3",
-    "uuid": "9f907d9c-a460-4a4e-a9e1-e79dfe91737e",
-    "type": "textfield",
-    "properties": {
-      "label": "Which population are you measuring in the numerator?",
-      "hint": "For example: The number of children enrolled in CHIP in the last federal fiscal year.",
-      "name": "goal_numerator_definition",
-      "multiline": true,
-      "title": "Define your numerator you’re measuring"
-    }
-  },  
-  {
-    "id": "4",  
-    "uuid": "c8f93c92-1d1d-4cf5-b1ca-b54cb659ed05",
-    "type": "textfield",
-    "properties": {
-      "label": "Numerator (Total number):",
-      "name": "goal_numerator",
-      "size": "small",
-      "numeric": true
-    }
-  },
-  {
-    "id": "5",
-    "uuid": "8cc86611-a4b5-4097-9f48-6cce5f563bbc",
-    "type": "textfield",
-    "properties": {
-      "label": "Which population are you measuring in the denominator?",
-      "hint": "For example: The total number of eligible children in the last federal fiscal year.",
-      "name": "goal_denominator_definition",
-      "multiline": true,
-      "title": "Define your denominator you’re measuring"
-    }
-  },
-  {
-    "id": "6",
-    "uuid": "60e05629-a0c0-4a5d-8026-ce42952c78d8",
-    "type": "textfield",
-    "properties": {
-      "label": "Denominator (Total number)",
-      "name": "goal_denominator",
-      "size": "small",
-      "numeric": true
-    }
-  },
-  {
-    "id": "7",
-    "uuid": "484c1190-68d5-4249-b46d-919bf5cec41f",
-    "type": "calculator",
-    "properties": {
-      "label": "Percentage",
-      "hint": "Auto-calculated",
-      "name": "goal_percentage"
-    }
-  },
-  {
-    "id": "8", 
-    "uuid": "55c4f5be-dfd1-41ea-98ef-e64f1322c6e2",
-    "type": "daterange",
-    "properties": {
-      "label": "What is the range",
-      "name": "goal_date_range",
-      "children": [ 
-        { 
-          "type": "datefield",
-          "properties": {
-              "label": "Start", 
-              "name": "goal_date_range_start",
-              "hint": "From mm/yyyy to mm/yyyy",
-              "month": true,
-              "day": false,
-              "year": true
-          }
-        },
-        {
-          "type": "datefield",
-          "properties": {
-            "label": "End", 
-            "name": "goal_date_range_end",
-            "hint": "From mm/yyyy to mm/yyyy",
-            "month": true,
-            "day": false,
-            "year": true
-          }
-        }
-      ]
-    }
-  },
-  {
-    "id": "9",
-    "uuid": "c586ad9d-e550-4eea-966a-fb618bb8d089",
-    "type": "choicelist",
-    "properties": {
-      "label": "Which data source did you use?",
-      "name": "goal_data_source",
-      "children": [
-        {
-          "defaultChecked": true,
-          "label": "Eligibility or enrollment data",
-          "value": "goal_enrollment_data"
-        },
-        {
-          "label": "Survey data", 
-          "value": "goal_survey_data" 
-        },
-        {
-          "label": "Another data source", 
-          "value": "goal_other_data"
-        }
-      ]
-    }
-  },
-  {
-    "id": "10",
-    "uuid": "9de55e16-b518-403d-a549-5923d740b7dd",
-    "type": "textfield",
-    "properties": {
-      "label": "How did your progress towards your goal last year compare to your previous year’s progress?",
-      "name": "goal_progress_comparison",
-      "multiline": true
-    }
-  },
-  {
-    "id": "11",
-    "uuid": "a21729a8-67ae-4e8e-9727-30b759624fa2",
-    "type": "textfield",
-    "properties": {
-      "label": "What are you doing to continually make progress towards your goal?",
-      "name": "goal_progress_action",
-      "multiline": true
-    }
-  },
-  {
-    "id": "12",
-    "uuid": "73972892-bfbf-42c4-a38c-1f4de9cbca1d",
-    "type": "textfield",
-    "properties": {
-      "label": "Anything else you'd like to tell us about this goal?",
-      "name": "goal_additional_information",
-      "multiline": true
-    }
-  },
-  {
-    "id": "13",
-    "uuid": "71d1ca8f9-8a85-425c-9feb-d9399074df11",
-    "type": "upload",
-    "properties": {
-      "label": "Do you have any supporting documentation? (Optional)",
-      "name": "goal_supporting_documentation",
-      "buttonLabel": "Browse"
-    }
-  }
-]
